replace a broken link in 2021-07-19-pyrse-book.md
merely-useful.tech has been abandoned; the book is now hosted elsewhere. Worth updating as this article is the first link when searching for the book on DuckDuckGo to mention the free online version.

The R link is also dead, but that project has morphed into something rather different to "a corresponding book for R users" and is now being coordinated by a different organisation, so that link no longer has a corresponding live version.